curch-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
beautifulsoup4==4.12.2
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
fastapi==0.110.0
google-search-results==2.4.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-community>=0.0.10
langchain-core==0.3.50
langchain-text-splitters==0.3.7
langsmith==0.3.23
orjson==3.10.16
packaging==24.2
python-dotenv==1.0.1
pydantic==2.11.1
pydantic_core==2.33.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.27.1
zstandard==0.23.0
